+++ packages/react/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @ponder/react
 
+## 0.12.23
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ponder/client@0.12.23
+
+## 0.12.22
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ponder/client@0.12.22
+
+## 0.12.21
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ponder/client@0.12.21
+
 ## 0.12.20
 
 ### Patch Changes